arch_util: move physical types utilites to their own file
